+++ src/__tests__/VariableSizeList.js
@@ -197,6 +197,23 @@ describe('VariableSizeList', () => {
       rendered.getInstance().scrollToItem(19, 'center');
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
     });
+
+    it('should scroll to the correct item for align = "smart"', () => {
+      const onItemsRendered = jest.fn();
+      const rendered = ReactTestRenderer.create(
+        <VariableSizeList {...defaultProps} onItemsRendered={onItemsRendered} />
+      );
+      // Scroll down enough to show item 10 in the middle.
+      rendered.getInstance().scrollToItem(10, 'smart');
+      // Scrolldn't scroll at all because it's close enough.
+      rendered.getInstance().scrollToItem(9, 'smart');
+      // Should scroll but not center because it's close enough.
+      rendered.getInstance().scrollToItem(6, 'smart');
+      // Item 1 can't align in the middle because it's too close to the beginning.
+      // Scroll up as far as possible though.
+      rendered.getInstance().scrollToItem(1, 'smart');
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
+    });
   });
 
   describe('resetAfterIndex method', () => {